requirements.txt
asgiref==3.2.10
Django==3.1
django-taggit==1.3.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
